requirements.txt
gspread
google-auth
google-auth-oauthlib
pandas
matplotlib
seaborn
openpyxl